----------------- -- Description -- ----------------- The Problem   When taking photos, the camera generates both raw (.CR2) and .JPG files. The   raw files are much larger than the JPG files. Viewers run much faster   rendering the JPG files than the raw files, so, I tend to view JPG files.   While viewing the files, I use the viewer to delete some of the JPG images.   Don't want to keep all the images captured. When a JPG is deleted, the   corresponding raw files should also be deleted.
This script examines both the jpg and the raw directory. Whenever a raw (.CR2)
file exists, but a corresponding .JPG does not exist, generate a rm() command
to remove the raw file. This is far easier, and FAR less error prone, than me
manually deleting the pair of raw and JPG files. Pipe the output into a file,
give it execute permission, and the raw files are easily and reliably deleted.

First functioning version.
CWD must be a "raw" directory. Will change in a future version.